requirements.txt
fastapi
uvicorn
PyMuPDF
tiktoken
python-multipart
nltk
llmsherpa
python-dotenv
openai
Pillow